// src/lib/enums/preferences/ServerPreferencesEnum.ts
export enum ServerPreferencesEnum {
  HTTP_PORT = 'SERVER.HTTP_PORT',
  HTTPS_PORT = 'SERVER.HTTPS_PORT',
  USE_HTTPS = 'SERVER.USE_HTTPS',
  PRIVATE_KEY = 'SERVER.PRIVATE_KEY',
  CERTIFICATE = 'SERVER.CERTIFICATE',
  GRAPHQL_ENDPOINT = 'SERVER.GRAPHQL_ENDPOINT',
  GRAPHIQL_ENDPOINT = 'SERVER.GRAPHIQL_ENDPOINT',
  USE_GRAPHIQL = 'SERVER.GRAPHIQL_ACTIVE',
  REST_ENDPOINT = 'SERVER.REST_ENDPOINT',
}